♻️ refactoring userdash
